streamlit/requirements.txt
streamlit==1.27.2
requests
psycopg2-binary